Code Cleanup (Indentation and Cast adjustments)
- Unsure if this may break anything, but needed to be done.

+++ src/main/java/net/id/aether/mixin/client/render/LivingEntityRendererMixin.java
@@ -13,9 +13,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(LivingEntityRenderer.class)
 public class LivingEntityRendererMixin {
     @Inject(method = "setupTransforms", at = @At("TAIL"))
-    private void setupTransforms(LivingEntity entity, MatrixStack matrices, float animationProgress, float bodyYaw, float tickDelta, CallbackInfo ci){
-        if(((AetherEntityExtensions)entity).getFlipped()) {
-            matrices.translate(0.0D, (double) (entity.getHeight() + 0.1F), 0.0D);
+    private void setupTransforms(LivingEntity entity, MatrixStack matrices, float animationProgress, float bodyYaw, float tickDelta, CallbackInfo ci) {
+        if (((AetherEntityExtensions) entity).getFlipped()) {
+            matrices.translate(0.0D, entity.getHeight() + 0.1F, 0.0D);
             matrices.multiply(Vec3f.POSITIVE_Z.getDegreesQuaternion(180.0F));
         }
     }
